(wip) Improve help output.
Add a real usage example, and remove list of templates.

We don't need to list available templates in the help output because we
have a command for that, and it will clutter up help as the list grows.

Note: this commit breaks the cucumber features. Expect this branch to be
rebased before merging.